fix: refactor client to fix dynamic imports
Fixes #3140.
