+++ client/src/features/ProjectPageV2/ProjectPageContent/ProjectPageContent.tsx
@@ -18,10 +18,10 @@
 import cx from "classnames";
 import { Col, Row } from "reactstrap";
 import { Project } from "../../projectsV2/api/projectV2.api.ts";
-import { ProjectPageContentType } from "../ProjectPageContainer/ProjectPageContainer.tsx";
 import ProjectInformation from "./ProjectInformation/ProjectInformation.tsx";
 import ProjectPageOverview from "./ProjectOverview/ProjectOverview.tsx";
 import ProjectPageSettings from "./Settings/ProjectSettings.tsx";
+import { ProjectPageContentType } from "./projectPageContentType.types.ts";
 
 export default function ProjectPageContent({
   project,
