+++ client/src/features/ProjectPageV2/ProjectPageContent/projectPageContentType.types.ts
@@ -13,24 +13,11 @@
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
- * limitations under the License
+ * limitations under the License.
  */
-import { Suspense, lazy } from "react";
-import PageLoader from "../../components/PageLoader";
-import { ProjectPageContentType } from "./ProjectPageContent/projectPageContentType.types";
 
-const ProjectPageV2Show = lazy(
-  () => import("./ProjectPageContainer/ProjectPageContainer.tsx")
-);
-
-export default function LazyProjectPageV2Show({
-  contentPage,
-}: {
-  contentPage: ProjectPageContentType;
-}) {
-  return (
-    <Suspense fallback={<PageLoader />}>
-      <ProjectPageV2Show contentPage={contentPage} />
-    </Suspense>
-  );
+export enum ProjectPageContentType {
+  Overview = "Overview",
+  Settings = "Settings",
+  ProjectInfo = "ProjectInfo",
 }
